pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools-scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
dynamic = ["version"]
name = "socks"
authors = [
    { name = "Marvin Fuchs", email = "marvin.fuchs@kit.edu" },
]
description = "A tool to build images for SoC devices."
readme = "README.md"
license = { file = "COPYING" }
requires-python = ">= 3.10"
dependencies = [
    "pyyaml>=5.4.1",
    "python-dateutil>=2.8.1",
    "requests>=2.25.1",
    "validators>=0.33.0",
    "tqdm>=4.57.0",
    "argcomplete>=3.5.0",
    "pydantic>=2.9.2"
]

[tool.setuptools.package-data]
socks = ["container/*.containerfile", "container/*.sh", "templates/**/*"]
amd_zynqmp_support = ["templates/**/*"]
amd_versal_support = ["templates/**/*"]
raspberrypi_support = ["templates/**/*"]

[tool.setuptools_scm]
version_scheme = "no-guess-dev"
local_scheme = "node-and-date"
relative_to = "__file__"
root = ".."

[tool.black]
line-length = 120

[project.scripts]
socks = "socks.main:main"
